This update brought to you by some Rachael induced insomnia...
Wrote generic SendRLPFrame support for fbus-3810/6110 and mbus-2110 code
thought the 6110 is the only one supported at present (just copied
what Pavels code in rlp-common seemed to be doing).  Will do the
corresponding 3810 code in the next day or two.  Judicious placement
of a return function in the rlp-common code means that it wont
use new function unless you remove same.

I've tried to keep the rlp-* code independant of the gsm-api code as
much as possible which has led to a slightly clumsy implementation I think.
Maybe it's OK really and I'm not thinking straight :)