Bug found and fixed by Ulrich Hopp:
Bug description :

When sending a SMS and choosing an addressee out of the phonebook or entering
a number with a corresponding phonebook entry I got the error message :

     Adress line contains illiegal adress.

Happens only with phonebook entries with leading or trailing white spaces,
spaces or tabs.

When xgnokii puts the phonebook entry to the adress line it "g_strstrip"s the
phonebook entry. ( => xgnokii_sms.c, line 766 )
( g_strstrip : a glib function which removes all leading and trailing white
spaces, spaces and tabs )
When the SMS should be sent, xgnokii looks for the corresponding number in
the phonebook by comparing all phonebook entries with the adress line.
Because in GUI_GetNumber ( xgnokii_contacts.c, line 3252 ) the phonebook
entry is not "g_strstrip"ped no match is found and the error message
appears.


Patch :

Although name is already "g_strstip"ped it is save to do it here again. I
think this makes it more "error resistant".