Drop "T" modifier from ATDT for compatibility and flexibility
Not all phones ignore it (eg. Sony Ericsson Z310i chokes on it) even if GSM
07.07 says its implementation is mandatory but it is ignored; also it
prevents the use of direct dialling from phonebooks because it conflicts
with the ">" modifier.
At this time the "ATD><str>" variant works only if encoding of <str> matches
what the phone expects, no conversion is done by libgnokii.