fix: workaround when gnokii hangs on opening the serial port
We've got reports that gnokii does hang when opening the serial port in AT
mode for Nokia phones. The debug output ends with the line:
 Serial device: opening device /dev/ttyACM0
The example reported device is:
 Bus 001 Device 002: ID 0421:047b Nokia Mobile Phones
Strace shows that the hang can be caused by:
 tcsetattr(fd, TCSADRAIN, &t);
or
 ioctl(fd, TIOCMBIS, &flags);
Apparently they go together (if one doesn't work the other does not work
neither). We suspect that this is the kernel problem. Some additional
reports to describe the problem symptoms:
http://old.nabble.com/Gnokii-hangs-on-ttyUSB0-at-an-ioctl-td7613918.html
https://bugzilla.redhat.com/show_bug.cgi?id=504798

The patch allows to disable setting port speed, DTR and RTS bits.
To disable setting port speed set:
 serial_baudrate = 0
in the config file. To disable DTS and RTS bits set:
 set_dtr_rts = 0
in the config file.

Thanks to Sapote at #gnokii IRC channel for help in debugging.