smsd: do not call GetSMSCenter every time SMS is sent
GetSMSC is not quite cheap operation. We do in smsd following:
  calloc()
  GetSMSC()
  strcpy()
  free()

On the other hand in 99.999% cases SMSC does not change often or does not
change at all.  It is safe to get the value once and then cache it for the
subsequent runs.

If GetSMSCenter fails do not attempt retries.