Do not make GN_PHONEBOOK_NUMBER_General the same as entry->number
Remove ambiguity about the number stored in entry->number (it was
either "TEL;TYPE=PREF,VOICE:" or "TEL;TYPE=PREF:" depending on the
order of appearance in the vcard file) and avoid duplication of the
General number, also output entry->number only when it is not empty.
This way the representation of the vcard in memory avoids loss of
information and when writing it back to file for testing we obtain a
file identical to the one read in the beginning.