Optional authentication for AT driver
These changes support devices that need authentication process when
initializing.  Support is provided just for AT driver.

The following methods of authentication are supported.

text
It is used for regular PIN/PUK authentication (ascii text data). Security
code will be read from the file or from the interactive prompt.  The file
format looks as follows:
 PIN:1234
 PUK:12345678
 PIN2:5678
 PUK2:0987654321
 SEC:12345
gnokii will issue:
 AT+CPIN?
Depending on the outcome different actions will be taken.
 +CPIN: READY - no action
 +CPIN: SIM PIN - authenticate using PIN
 +CPIN: SIM PUK - authenticate using PUK
 +CPIN: SIM PIN2 - authenticate using PIN2
 +CPIN: SIM PUK2 - authenticate using PUK2

interactive
It allows program that uses libgnokii to register own authentication code.
The indented use case is for interactive entering PIN, PUK or other security
code.  The registered function should determine on its own whether any
security code is expected and if so, which one.

binary
For future use. Intended usage is for any binary data authentication.

Authentication type and path to the relevant file are stored in gnokii
config.  It can be used either in [global] section or in [phone_*] sections.