Do not ignore the validity value passed to sms_prepare()
The default value is still 3 days so nothing changes for users not
explicitely setting a --validity value, other users will see their
value applied.
The value for TP-Validity-Period (TP-VP) set by the user was
overwritten: first sms->validity was set in sms_default() as number of
minutes corresponding to 3 days, then the value could be overridden by
a program (eg. gnokii --sendsms --validity n), setting a possibly
different number of minutes, but sms_prepare() would ignore it and set
rawsms->validity[0] to a constant encoded value meaning again "3 days",
so that changing the default value was in fact impossible.