Avoid buffer overrun for the main phone number
Use the correct size for the main phone number which has less space than
subentries.
This went unnoticed because in the test vcard the tail of the long
TEL;TYPE=PREF,VOICE number was overwritten by values coming later in the
file and a zero byte in those values was likely to provide a string terminator.