Reading 0 bytes is a simpler way to send an empty SMS
Empty SMS are allowed on purpose as requested by an user since commit
8c4f263c6c8b99d042231caf3006b95d54128a99
This also removes one translatable string.
After this change this is possible:
gnokii --sendsms +123456789 </dev/null
Before this change the only way to send an empty SMS was to give an
empty line:
echo | gnokii --sendsms +123456789
A trailing \n is still stripped by readtext() so this still works.